Move BlockValue into block_handler module
This is a better fit since they belong together.
