+++ src/lib.rs
@@ -132,7 +132,6 @@ extern crate alloc;
 pub mod error;
 
 pub mod block_handler;
-pub mod block_value;
 mod header;
 pub mod link_format;
 #[macro_use]
